turn into generic

+++ Entities/Generation.cs
@@ -2,7 +2,7 @@ namespace FroniusIntegration.Entities;
 
 public class Generation
 {
-  public Generation(int? locationId, decimal? volumeKwh, DateTime createAt, DateTime updateAt)
+  public Generation(int? locationId, float? volumeKwh, DateTime createAt, DateTime updateAt)
   {
     LocationId = locationId;
     VolumeKwh = volumeKwh;
@@ -11,7 +11,7 @@ public class Generation
   }
 
   public int? LocationId { get; set; }
-  public decimal? VolumeKwh { get; set; }
+  public float? VolumeKwh { get; set; }
   public DateTime CreateAt { get; set; }
   public DateTime UpdateAt { get; set; }
 }
